+++ slacktools/api/web/pins.py
@@ -5,13 +5,15 @@ from typing import (
 
 from slacktools.api.base import BaseApiObject
 
+from .types import PinApiObjectType
+
 
 class PinApiMessageItem:
     type: str   # 'message'
     text: str
     user: str   # user id / author
     ts: str
-    blocks: Dict
+    blocks: List[Dict]
     pinned_to: List[str]  # List[channel_id]
     permalink: str
 
@@ -23,5 +25,5 @@ class PinApiObject(BaseApiObject):
     channel: str     # channel id
     message: PinApiMessageItem
 
-    def __init__(self, event_dict):
+    def __init__(self, event_dict: PinApiObjectType):
         super().__init__(event_dict=event_dict)